添加鉴权模块相关代码

+++ api-common/src/main/java/com/fdzang/microservice/common/entity/auth/AuthResult.java
@@ -8,7 +8,7 @@ import lombok.Data;
  */
 @Data
 public class AuthResult {
-    private int status;
+    private long status;
     private String description;
     private String orgCode;
     private String serviceName;
